ml/requirements.txt
speechrecognition
elevenlabs
python-dotenv
pyaudio
